refactor(mcp): split search.py into focused modules and migrate to pydantic.JsonValue
